src/requirements.txt
numpy==1.12.1
scipy==0.18.1
tensorflow_gpu==1.4.1
toml==0.9.2
sox==1.3.3
